requirements.txt
torch
torchvision
opencv-python
numpy
efficientnet_pytorch
requests
pillow
